+++ resolwe/flow/models/base.py
@@ -101,17 +101,6 @@ class BaseQuerySet(PermissionQuerySet):
 class BaseManager(PermissionManager[M, Q]):
     """Base manager set for Resolwe's ORM objects."""
 
-    # The data is partitioned into groups with the same partition fields value. The
-    # object with the greatest version inside the partition is considered the latest.
-    partition_fields: list[str] = ["slug"]
-
-    # The field to use as version. The latest version is determined by sorting the
-    # data by version field in the ascending order and taking the first entry.
-    version_field: str = "version"
-
-    # The created field is used to determine the version at certain point in time.
-    created_field: str = "created"
-
     # Versioning is enabled by default.
     enable_versioning: bool = True
 
@@ -127,8 +116,8 @@ class BaseManager(PermissionManager[M, Q]):
         latest_entries = queryset.annotate(
             rank=models.Window(
                 expression=models.functions.DenseRank(),
-                partition_by=[models.F(field) for field in self.partition_fields],
-                order_by=models.F(self.version_field).desc(),
+                partition_by=[models.F(field) for field in self.model.partition_fields],
+                order_by=models.F(self.model.version_field).desc(),
             ),
         ).filter(rank=1)
         return latest_entries
@@ -144,7 +133,7 @@ class BaseManager(PermissionManager[M, Q]):
 
     def history(self, timestamp: datetime.datetime) -> Q:
         """Get values at certain point in time."""
-        created_filter = {f"{self.created_field}__le": timestamp}
+        created_filter = {f"{self.model.created_field}__le": timestamp}
         filtered_queryset: Q = self.filter(**created_filter)  # type: ignore
         return self._only_latest(filtered_queryset)
 
@@ -180,6 +169,17 @@ class UniqueSlugError(IntegrityError):
 
 class BaseModel(AuditModel):
     """Abstract model that includes common fields for other models."""
+
+    # The data is partitioned into groups with the same partition fields value. The
+    # object with the greatest version inside the partition is considered the latest.
+    partition_fields: list[str] = ["slug"]
+
+    # The field to use as version. The latest version is determined by sorting the
+    # data by version field in the ascending order and taking the first entry.
+    version_field: str = "version"
+
+    # The created field is used to determine the version at certain point in time.
+    created_field: str = "created"
 
     class Meta:
         """BaseModel Meta options."""